OE0-76 user districts return type ≠ on tech users
The return for InteractiveUser was modified to be a queryset but the exception for other user kinds returned a list. So calls to values_list() on it failed.
We are now returning an empty queryset of the right type and we add a warning when called with a technical user as the process should not even have reached the list of districts for such a user.