Fix performance on user districts
The get_user_districts was trying to be more efficient with a .only() but not fetching enough fields, triggering a n+1 query. Should also have done a prefetch_related instead of select_related (not much effect here though).
In locations like DRC, this brings the user districts graphql call from 12sec to a few milliseconds !